Add job for swift-format (#3)
NeverForceUnwrap and NeverUseForceTry were removed. A CLI is run
synchronously to exit immediately such that a crash is better than
silently dropping the error and moving on.

AlwaysUseLowerCamelCase is ignored in a few files in the Tests folder
because test function names like compactMap_map are better than
compactMapMap.

Ignoring a rule in a file works since XCode 16.3 which is only available
in the macOS-15 runner image. For consistency, the versions were also
bumped in the test-macos job.

For more details, see the discussion in the PR.

See also: https://github.com/md-babel/swift-markdown-babel/pull/3
Related to: https://github.com/md-babel/swift-markdown-babel/issues/1